pyproject.toml
[tool.poetry]
name = "dnr-gnn"
version = "0.1.0"
description = ""
authors = ["Dennis Hollanders <dennis.hollanders17@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
folium = "^0.18.0"
networkx = "^3.4.2"
geopandas = "^1.0.1"
matplotlib = "^3.9.2"
shapely = "^2.0.6"
pyshp = "^2.3.1"
numpy = "<2.0"
ipykernel = "^6.29.5"
jupyter = "^1.1.1"
scipy = "<1.14"
plotly = "^5.24.1"
pandapower = "2.14.11"
fiona = "^1.10.1"
pyarrow = "^18.1.0"
torch = "^2.5.1"
torch-geometric = "^2.6.1"
scikit-learn = "1.3.2"
wandb = "^0.19.2"
lightning = "^2.5.0.post0"
gurobipy = "11.0.3"
lxml = "5.2.2"
pyomo = {extras = ["quadratic"], version = "^6.9.2"}
seaborn = "^0.13.2"
python-dotenv = "^1.0.1"
julia = "^0.6.2"
numba = "^0.61.0"
igraph = "^0.11.8"
simbench = "^1.5.3"
graphviz = "^0.20.3"
optuna = "^4.2.1"
optuna-integration = {extras = ["wandb"], version = "^4.2.1"}
psutil = "^7.0.0"
pip = "^25.0.1"
pyyaml = "^6.0.2"
pandas = "^2.2.3"
jellyfish = "^1.1.3"
python-levenshtein = "^0.27.1"
openpyxl = "^3.1.5"
fuzzywuzzy = "^0.18.0"
torchviz = "^0.0.3"
pint = "^0.24.4"
pulp = "^3.1.1"
cvxpy = "^1.6.4"
cvxpylayers = "^0.1.9"
typing-extensions = "^4.13.1"
ecos = "^2.0.14"
tensorboard = "^2.19.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
